Move cql_fixtures
