+++ lib/elasticsearch/cql/index-mapping.js
@@ -112,7 +112,7 @@ const indexMapping = {
     ]
   },
   callnumber: {
-    fields: ['shelfMark'],
+    fields: ['shelfMark', 'items.shelfMark'],
     term: ['shelfMark.keywordLowercased', 'items.shelfMark.keywordLowercased']
   },
   identifier: {
